+++ toshi_hazard_store/model/hazard_models_pydantic.py
@@ -1,10 +1,14 @@
 """The hazard metatdata models for (de)serialisation as json."""
 
 from datetime import datetime, timezone
+from typing import List
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, field_validator
 
 from toshi_hazard_store.oq_import.aws_ecr_docker_image import AwsEcrImage
+
+# Float32 = Annotated[float, {"pyarrow_type": pyarrow.float32()}]
+# Int32 = Annotated[int, {"pyarrow_type": pyarrow.int32()}]
 
 
 class CompatibleHazardCalculation(BaseModel):
@@ -55,3 +59,34 @@ class HazardCurveProducerConfig(BaseModel):
         """The unique ID should not include any non cross-platform characters (for filename compatablity)."""
         assert self.ecr_image_digest[:7] == "sha256:"
         return self.ecr_image_digest[7:]
+
+
+class HazardAggregateCurve(BaseModel):
+    """Generally these are aggregated realisation curves.
+
+    Attributes:
+        compatible_calc_id: for hazard-calc equivalence.
+        model_id: the model that these curves represent.
+        nloc_001:  the location string to three places e.g. "-38.330~17.550".
+        nloc_0:  the location string to zero places e.g.  "-38.0~17.0" (used for partitioning).
+        imt: the imt label e.g. 'PGA', 'SA(5.0)'
+        vs30: the VS30 integer
+        aggr: the aggregation type
+        values: a list of the 44 IMTL values
+    """
+
+    compatible_calc_id: str
+    hazard_model_id: str
+    nloc_001: str
+    nloc_0: str
+    imt: str
+    vs30: int
+    aggr: str
+    values: List[float]
+
+    @field_validator('values', mode='before')
+    @classmethod
+    def validate_len_values(cls, value: List) -> List:
+        if not len(value) == 44:
+            raise ValueError(f'expected 44 values but there are {len(value)}')
+        return value