+++ toshi_hazard_store/scripts/ths_ds_sanity.py
@@ -34,7 +34,7 @@ import toshi_hazard_store.config
 import toshi_hazard_store.model.openquake_models
 import toshi_hazard_store.model.revision_4.hazard_models  # noqa: E402
 import toshi_hazard_store.query.hazard_query
-from toshi_hazard_store.model.revision_4 import pyarrow_dataset
+from toshi_hazard_store.model.pyarrow import pyarrow_dataset
 from toshi_hazard_store.oq_import.oq_manipulate_hdf5 import migrate_nshm_uncertainty_string
 from toshi_hazard_store.scripts.core import echo_settings  # noqa